backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
httpx==0.27.0
openai==1.30.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
